// src/lib/types.ts
export interface TimeTrackerOverTime {
  isOverTime: boolean;
  reason: string | null;
}

export interface TimeTracker {
  _id: string;
  overTime: TimeTrackerOverTime;
  totalTimeSpend: number;
  project: TimeTrackerProject;
  maximumTimeSeconds: number;
  budget?: number;
  earnings?: number;
  hourlyRate?: number;
  entryType?: string;
  manualProjectName?: string;
}

export interface TimeTrackerState {
  _id: string;
  totalDuration: number;
  status: "inProgress" | "completed" | "paused" | "done" | "pause";
  tag: string;
  note?: string;
}

export interface TimeTrackerSession {
  _id: string;
  startTime: string;
  endTime: string | null;
  duration: number;
  task: string;
  createdAt: string;
  updatedAt: string;
}
export interface AdminTimeTracker {
  _id: string;
  overTime: TimeTrackerOverTime;
  hourlyRate: number;
  budget: number;
  earnings: number;
  totalTimeSpend: number;
  maximumTimeSeconds: number;
  projectName: string;
  manualProjectName?: string;
  entryType?: string;
  designer: string;
  customer?: string;
}